requirement.txt
openai
langchain
streamlit
google-generativeai
